# Optimized MLP with W/H fracturing, sharding and ReduceScatter