--- a/models/demos/t3000/falcon40b/tests/ci/test_falcon_end_to_end_60_layer_t3000_prefill_10_loops.py
+++ b/models/demos/t3000/falcon40b/tests/ci/test_falcon_end_to_end_60_layer_t3000_prefill_10_loops.py
@@ -96,35 +96,35 @@ def test_FalconCausalLM_prefill_end_to_end_t3000_ci_loops_10(
 
     if data_type == "BFLOAT8_B":
         if seq_len == 32:
-            out_pcc = 0.984
-            k_cache_pcc = 0.979
-            v_cache_pcc = 0.940
-            token_pcc = 0.99
-        elif seq_len == 128:
-            out_pcc = 0.989
-            k_cache_pcc = 0.989
-            v_cache_pcc = 0.949
-            token_pcc = 0.99
-        elif seq_len == 2048:
-            out_pcc = 0.993
-            k_cache_pcc = 0.991
-            v_cache_pcc = 0.97
-            token_pcc = 0.99
-    elif data_type == "BFLOAT16":
-        if seq_len == 32:
             out_pcc = 0.986
-            k_cache_pcc = 0.993
-            v_cache_pcc = 0.978
+            k_cache_pcc = 0.978
+            v_cache_pcc = 0.934
             token_pcc = 0.99
         elif seq_len == 128:
-            out_pcc = 0.991
-            k_cache_pcc = 0.994
-            v_cache_pcc = 0.982
+            out_pcc = 0.990
+            k_cache_pcc = 0.988
+            v_cache_pcc = 0.940
             token_pcc = 0.99
         elif seq_len == 2048:
             out_pcc = 0.992
-            k_cache_pcc = 0.992
-            v_cache_pcc = 0.980
+            k_cache_pcc = 0.990
+            v_cache_pcc = 0.967
+            token_pcc = 0.99
+    elif data_type == "BFLOAT16":
+        if seq_len == 32:
+            out_pcc = 0.981
+            k_cache_pcc = 0.978
+            v_cache_pcc = 0.929
+            token_pcc = 0.99
+        elif seq_len == 128:
+            out_pcc = 0.991
+            k_cache_pcc = 0.993
+            v_cache_pcc = 0.976
+            token_pcc = 0.99
+        elif seq_len == 2048:
+            out_pcc = 0.992
+            k_cache_pcc = 0.989
+            v_cache_pcc = 0.972
             token_pcc = 0.99
 
     disable_persistent_kernel_cache()