Optimized falcon40b MLP
- Use 2D weight stationary fracturing (FF1 in W and FF2 in H)
- MLP in two slices and sharded
- ReduceScatter after FF2 instead of AllGather between FF1 and FF2
- Also added for decode (different sharding scheme, otherwise same concept)

Perf
S=128: 1456 -> 1859 tok/s
S=2048: 2857 -> 3355 tok/s

CI
- [x] Frequent tests: https://github.com/tenstorrent/tt-metal/actions/runs/9991004079
- [x] Unit tests: https://github.com/tenstorrent/tt-metal/actions/runs/9991008499
- [x] Demo tests: https://github.com/tenstorrent/tt-metal/actions/runs/9990971375
- [x] Perf tests: https://github.com/tenstorrent/tt-metal/actions/runs/9991012788